combine engine reader and writer to chunking module
