+++ src/kanta/engine/chunking.py
@@ -1,11 +1,35 @@
 import shutil
+from collections.abc import Iterator
 from pathlib import Path
 
 import pandas as pd
 import pyarrow.parquet as pq
 
+from kanta.config import ENGINE_CHUNK_N_LINES, ENGINE_INPUT_COLUMNS_MAPPING
+
 CHUNKS_FILE_TEMPLATE = "chunk_{index:06d}.parquet"
 CHUNKS_FILE_GLOB = "chunk_*.parquet"
+
+
+def chunk_iterator(
+    input_file: Path, *, is_test_run: bool = False
+) -> Iterator[pd.DataFrame]:
+    # Use pyarrow to read the Parquet file in chunks.
+    parquet_file = pq.ParquetFile(input_file)
+
+    for batch in parquet_file.iter_batches(
+        batch_size=ENGINE_CHUNK_N_LINES,
+        # Select only the given columns, this speeds up the read quite a lot for Parquet files.
+        columns=ENGINE_INPUT_COLUMNS_MAPPING.keys(),
+    ):
+        yield batch.to_pandas(
+            # Use nullable data-types backed by pyarrow.
+            types_mapper=pd.ArrowDtype
+        )
+
+        # Yield only the first item when test run is ON
+        if is_test_run:
+            break
 
 
 def write_chunk(dataframe: pd.DataFrame, chunks_dir: Path, chunk_index: int) -> Path:
